AP_EFI_Hirth : timeout implementation

+++ libraries/AP_EFI/AP_EFI_Serial_Hirth.h
@@ -94,7 +94,7 @@ private:
     data_set_t res_data;
 
     //
-    uint8_t num_bytes;
+    uint32_t num_bytes;
 
     //
     // uint8_t computed_checksum;
